use existing aget_config to get dynamic config; add missing support for disabled transfer feature flag in miner

+++ miner/app/src/compute_horde_miner/miner/eviction.py
@@ -3,6 +3,7 @@ from datetime import timedelta
 
 from compute_horde.receipts.models import JobAcceptedReceipt, JobFinishedReceipt, JobStartedReceipt
 from compute_horde.receipts.store.local import LocalFilesystemPagedReceiptStore
+from compute_horde_miner.miner.receipts import current_store
 from django.utils.timezone import now
 
 from .models import AcceptedJob
@@ -31,6 +32,7 @@ def evict_receipts() -> None:
     JobAcceptedReceipt.objects.filter(timestamp__lt=cutoff).delete()
     JobFinishedReceipt.objects.filter(timestamp__lt=cutoff).delete()
 
-    logger.info("Evicting old receipt pages")
-    store = LocalFilesystemPagedReceiptStore()
-    store.delete_pages_older_than(cutoff)
+    store = current_store()
+    if isinstance(store, LocalFilesystemPagedReceiptStore):
+        logger.info("Evicting old receipt pages")
+        store.delete_pages_older_than(cutoff)
